bubble sort

+++ 0-bubble_sort.c
@@ -10,7 +10,7 @@ void bubble_sort(int *array, size_t size)
 {
 	size_t i;
 	size_t j;
-	size_t temp;
+	int temp;
 
 	if (array == NULL || size < 2)
 		return;
